% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bound_analysis.R
\name{lower_bound_analysis}
\alias{lower_bound_analysis}
\title{Run model experiments.}
\usage{
lower_bound_analysis()
}
\description{
This runs the experiments where data sets are simulated, and a basic
capture-recapture model (M0) is compared against a model that includes
encounter data leading to failed captures.
}
