% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m0_experiments.R
\name{m0_experiments}
\alias{m0_experiments}
\title{Run M0 model experiments.}
\usage{
m0_experiments()
}
\description{
This runs the experiments where data sets are simulated, and a basic
capture-recapture model (M0) is compared against a model that includes
encounter data leading to failed captures.
}
